+++ src/main/java/core/Currency/TransmutationOrb.java
@@ -8,7 +8,7 @@ import core.Utils.AddRandomMod;
 
 public class TransmutationOrb implements Crafting_Action {
 
-    private double cost = 1.0;
+    private int cost = 1;
 
     public enum CurrencyTier {
         BASE, GREATER, PERFECT
@@ -63,7 +63,7 @@ public class TransmutationOrb implements Crafting_Action {
     }
 
     @Override
-    public double getCost() {
+    public int getCost() {
         return cost;
     }
